refactor lib compilation
