+++ .travis.yml
@@ -12,6 +12,7 @@ install:
   - npm run setup
 
 script:
+  - npm run clean
   - npm run buildWebpack
   - npm run unitall -- -w 1
   - npm run tslint
