+++ package.json
@@ -10,8 +10,8 @@
         "reset": "rmdir /S /Q node_modules",
         "dev": "rimraf dist/**/* && webpack-dev-server --config conf/webpack.config.js --content-base src --colors --port 8082 --open --history-api-fallback",
         "buildWebpack": "webpack --config conf/webpack.config.js --colors --env.buildonly",
-        "build": "rimraf dist/**/* && tsc && copyfiles -u 1 \"./src/**/*.json\" dist && copyfiles -u 1 \"./src/**/*.html\" dist && copyfiles -u 1 \"./src/**/*.scss\" dist && copyfiles -u 1 \"./src/**/*.css\" dist && npm run assetsToDist && npm run metagen",
-        "buildlib": "rimraf dist/**/* && webpack --config conf/webpack.config.js --colors --env.lib",
+        "build": "tsc && copyfiles -u 1 \"./src/**/*.json\" dist && copyfiles -u 1 \"./src/**/*.html\" dist && copyfiles -u 1 \"./src/**/*.scss\" dist && copyfiles -u 1 \"./src/**/*.css\" dist && npm run assetsToDist && npm run metagen",
+        "buildlib": "webpack --config conf/webpack.config.js --colors --env.lib",
         "dtsToDist": "copyfiles -f \"./src/**/*.d.ts\" dist",
         "scssToDist": "copyfiles -u 1 \"./src/styles/**/*.scss\" dist && copyfiles -u 1 \"./src/styles/**/*.css\" dist",
         "assetsToDist": "copyfiles -u 1 \"./src/assets/**/*\" dist",
