+++ tsconfig.dev.json
@@ -20,6 +20,8 @@
     },
     "exclude": [
         "src/**/*.spec.*",
+        "dist/**/*",
+        "lib/**/*",
         "tests/**/*"
     ]
 }
